refactor(library): split history screen and calendar for detekt limits
Extract history top bar, body, dialogs, insight carousel, and activity
graphs so HistoryScreen/HistoryCalendar stay under LongMethod and
TooManyFunctions without baseline changes.
